requirements.txt
PyYAML~=6.0
GDAL~=3.4.1
laspy~=2.1.1
numpy~=1.22.2
setuptools~=60.8.1